removes github workflows
Instead of generating the SDK from a workflow in this directory, I am
instead opting to create a workflow where the OpenAPI spec is kept.
It'll be easier to trigger the workflow on changes to the spec as well
as re-use workflow components for generating SDK code for other
languages in the future.