brushing off this repo and recreating a TS API for Hessra
Creating a simple HessraClient to use the Hessra authorization service
API based on the openapi_spec.

The goal is to get this publishable to node and usable by nextjs and
maybe deno later on.

Future plans for this include using our parts of our rust-sdk for local
token verification and configuration as WASM modules in this to create a
full-fledged SDK for various typescript things.